requirements.txt
asgiref==3.6.0
Django==3.2.19
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
gunicorn==20.1.0
whitenoise==6.2.0